Add CFG visualizer
Since the book provides a logical schema diagram which corresponds to a CFG, I've decided to take advantage of that. The CFG visualizer outputs a PNG of labelled basic blocks. One operator corresponds to 1-or-more basic blocks, each named either `op. x` or `op x. + n` (representing an offset from the start of the operator).