Add documentation to the DSL helpers
To clarify the exact role of the DSL helpers, I've added documentations
to many of the most confusing instructions. I've also added a couple
functions that should help clarify the function of pieces of code like
shift. I've also added a note about which branch gets executed for comparisons
since that always confuses me (not a good sign :P).