Add "Operator Trace" debugging support
Instead of having to watch _every_ instruction, the
idea is to instead only look at the flow between basic-blocks
of the assembly. Using this it becomes easier to see patterns in the control
flow. To make this work, the upper 25 bits of the generated binary are used
to store the id of the basic block and procedure. The VM also has a special
mode that builds an ordered log of the blocks that were visited. This log
can be turned into an image using a new "trace" executable in this project.