Fix longstanding bug with global Size constants
A while ago I added support for "packing" Size constants in the disk
format since they are empty. However, this means that we need to adjust
the target address when performing an Ma/Mb instruction, to "skip" all
the packed cells. Well I finally realized what the problem was today and
made the assembler track the amount of cells that got packed so that
load addresses can be adjusted correspondingly. With this change PP-2
now properly loads! This means that I can finally make actual progress.